+++ CHANGELOG.md
@@ -1,8 +1,8 @@
-# Keystone Percentage Helper
+# Keystone Polaris
 
-## [2.1](https://github.com/ZelionGG/KeystonePercentageHelper/releases/tag/2.1) (2025-09-18)
+## [3.0](https://github.com/ZelionGG/KeystonePolaris/releases/tag/3.0) (2025-10-06)
 
-[Full Changelog](https://github.com/ZelionGG/KeystonePercentageHelper/compare/2.0.1...2.1) [Previous Releases](https://github.com/ZelionGG/KeystonePercentageHelper/releases)
+[Full Changelog](https://github.com/ZelionGG/KeystonePolaris/compare/2.0.1...3.0) [Previous Releases](https://github.com/ZelionGG/KeystonePolaris/releases)
 
 - Introduced a unified display with optional **Total required for section**, **Current percentage** and **Pull percentage** (Requires MDT), projected values (parentheses) in combat for Required/Current, per-line completion suffixes, configurable layout (single-line or multi-line), and a switch between percentage and count-based modes.
 - Russian translation updated, thank you **Hollicsh**.